requirements.txt
requests~=2.28.1
argparse~=1.4.0
flask~=2.1.3
textblob~=0.17.1
python-kasa
aioflask